+++ packages/riverpod_devtools/lib/src/http_server_stub.dart
@@ -1,0 +1,7 @@
+class RiverpodDevToolsHttpServer {
+  const RiverpodDevToolsHttpServer({int maxBufferSize = 1000});
+  Future<void> start() async {}
+  void stop() {}
+  void addEvent(Map<String, Object?> event) {}
+  void clearEvents() {}
+}